update to move current goals to separate js file
